chore: rename config to network
