+++ pkg/options/nats.go
@@ -5,13 +5,13 @@ import (
 	"log/slog"
 
 	"github.com/nats-io/nats.go"
-	"github.com/synternet/data-layer-sdk/pkg/config"
+	"github.com/synternet/data-layer-sdk/pkg/network"
 )
 
 // MakeNats is a convenience function to create NATS connection from various parameters.
 func MakeNats(name, urls, userCreds, nkey, jwt, caCert, clientCert, clientKey string) (*nats.Conn, error) {
 	if urls == "" {
-		urls = config.DefaultNetwork().JoinURLs()
+		urls = network.Default().JoinURLs()
 		if urls == "" {
 			slog.Info("Will use NATS stub")
 			return nil, nil
